# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  legocy-go_internal_delievery_http_resources_lego.LegoSeriesRequest:
    properties:
      name:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_lego.LegoSeriesResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_lego.LegoSetRequest:
    properties:
      n_pieces:
        type: integer
      name:
        type: string
      number:
        type: integer
      series_id:
        type: integer
    type: object
  legocy-go_internal_delievery_http_resources_lego.LegoSetResponse:
    properties:
      id:
        type: integer
      n_pieces:
        type: integer
      name:
        type: string
      number:
        type: integer
      series:
        $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSeriesResponse'
    type: object
...
  legocy-go_internal_delievery_http_resources_marketplace_admin.MarketItemAdminResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      lego_set:
        $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSetResponse'
      location:
        type: string
      price:
        type: number
      seller:
        $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users.UserDetailResponse'
      set_state:
        type: string
      status:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_marketplace_admin.MarketItemAdminUpdateRequest:
    properties:
      description:
        type: string
      lego_set_id:
        type: integer
      location:
        type: string
      price:
        type: number
      seller_id:
        type: integer
      set_state:
        type: string
      status:
        type: string
    type: object
...
  legocy-go_internal_delievery_http_resources_users.UserDetailResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: LEGOcy is a marketplace for LEGO lovers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LEGOcy API
  version: "1.0"
paths:
...
  /admin/market-items/{itemId}:
    get:
      operationId: detail_market_item_admin
      parameters:
      - description: item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_marketplace_admin.MarketItemAdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Get Market Item by ID (Admin)
      tags:
      - market_items_admin
    put:
      operationId: update_market_item_admin
      parameters:
      - description: item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_marketplace_admin.MarketItemAdminUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_marketplace_admin.MarketItemAdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Update Market Item (Admin)
      tags:
      - market_items_admin
  /admin/series/:
    post:
      operationId: create_series
      parameters:
      - description: create data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSeriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Create LEGO Series object
      tags:
      - lego_series_admin
  /admin/series/{seriesID}:
    delete:
      operationId: delete_series
      parameters:
      - description: series ID
        in: path
        name: seriesID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Delete LegoSeries object
      tags:
      - lego_series_admin
  /admin/sets/:
    post:
      operationId: set_create
      parameters:
      - description: create data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Create Lego Set object
      tags:
      - lego_sets_admin
  /admin/sets/{setID}:
    delete:
      operationId: set_delete
      parameters:
      - description: LegoSet ID
        in: path
        name: setID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Delete Lego Set object
      tags:
      - lego_sets_admin
...
  /series/:
    get:
      operationId: list_lego_series
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSeriesResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: List of LEGO Series objects
      tags:
      - lego_series
  /series/{seriesID}:
    get:
      operationId: detail_lego_series
      parameters:
      - description: series ID
        in: path
        name: seriesID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSeriesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: LEGO Series by id
      tags:
      - lego_series
...
  /sets/{setID}:
    get:
      operationId: set_detail
      parameters:
      - description: Lego Set ID
        in: path
        name: setID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Get LegoSet by ID
      tags:
      - lego_sets
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"